Test custom emoticons
This is almost an exact clone of the tinymce emoticons plugin. We want to do this so that we can add our own custom emoticons. Also, we want to customize the location of where these emoticons will be stored, so that we do not have to permanently keep the emoticons of every tinymce version we use in the CDN.