Add a key property that updates the displayed content when changed.
When the key property is changed, it likely means that the html-editor is being re-used (e.g. in a dom repeat) to represent a different entity.
We re-set the html-editor content to ensure the displayed value in the html editor and the value passed in are not out of sync.