Clear content not performing as expected
Fixes...
Issue 1: If we explicitly 'clearContent', we should clear the undo stack as well.
Issue 2: When we explicitly setContent, an undo level is not added. This means that before and after clearing, if the content is the same (i.e. paste the same content back in), the editor will think the content stayed the same, and will not fire a change event.
Issue 3: Explicitly setting the content does not fire a 'change' event. We should fire a change event so that consumers of the editor will be notified that it is cleared.